backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
python-dotenv==1.0.1
pydantic==2.6.4
pydantic-settings==2.2.1
google-generativeai==0.4.1
motor==3.3.2
pymongo==4.6.2
pypdf==4.1.0
